+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.7.0
 
+- Renames `Process.pipe_into` to `.pipe`.
 - Adds `Process.then` as an analogue of shell `P ; Q`.
 - Adds `Process.or_else` as an analogue of shell `P || Q`.
 - Adds `CompletedProcess.ok` as a return-boolean version of `.check`.